Replaced botlib with one from CNQ3

+++ code/botlib/l_struct.h
@@ -48,7 +48,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 typedef struct fielddef_s
 {
 	char *name;										//name of the field
-	int offset;										//offset in the structure
+	intptr_t offset;										//offset in the structure
 	int type;										//type of the field
 	//type specific fields
 	int maxarray;									//maximum array size
